+++ Test/Snapshots/edXTests.CourseOutlineHeaderViewTests/testCourseVideosHeaderView_ios10_380x72@2x.png
