+++ Test/Snapshots/edXTests.CourseOutlineHeaderViewTests/testCourseVideosHeaderView_ios9_380x72@2x.png
