+++ Test/Snapshots/edXTests.CourseOutlineHeaderViewTests/testCourseVideosHeaderView_ios9_rtl_380x72@2x.png
